Renamed several files so the new naming matches our new naming convention
