+++ docs/source/examples.rst
@@ -43,9 +43,9 @@ In this scenario, human drivers placed on a ring develop a travelling wave that 
 decelerations and lowers the average velocity of the system. The goal is to train a single
 autonomous vehicle to eliminate the shockwave.
 
-The pre-built run script for running this in human-only mode is `examples/sumo/sugiyama.py`.
-The run scripts in rllib and aimsun are `examples/rllib/stabilizing_the_ring.py` and
-`examples/aimsun/stabilizing_the_ring.py`. These control environments are partially observed:
+The experiment config for this in human-only mode is `examples/exp_configs/non_rl/ring.py`.
+The run scripts in rllib and aimsun are `examples/exp_configs/singleagent/singleagent_ring.py` 
+and `examples/aimsun/singleagent_ring.py`. These control environments are partially observed:
 the autonomous vehicle only observes its own distance to the leading vehicle, its speed,
 and the speed of the leading vehicle.
