+++ examples/train_rllib.py
@@ -38,7 +38,7 @@ def parse_args(args):
     parser.add_argument(
         'exp_config', type=str,
         help='Name of the experiment configuration file, as located in '
-             'exp_configs/single_agent.')
+             'exp_configs/singleagent.')
 
     return parser.parse_known_args(args)[0]
 
@@ -117,7 +117,7 @@ if __name__ == "__main__":
     flags = parse_args(sys.argv[1:])
 
     # import relevant information from the exp_config script
-    module = __import__("exp_configs.single_agent", fromlist=[flags.exp_config])
+    module = __import__("exp_configs.singleagent", fromlist=[flags.exp_config])
     submodule = getattr(module, flags.exp_config)
     flow_params = submodule.flow_params
     n_cpus = submodule.N_CPUS
