+++ tests/stress_tests/stress_test_start.py
@@ -28,13 +28,13 @@ def start():
     initial_config = InitialConfig(bunching=20)
 
     network = RingNetwork(
-        name="sugiyama",
+        name="ring",
         vehicles=vehicles,
         net_params=net_params,
         initial_config=initial_config)
 
     env = AccelEnv(env_params, sim_params, network)
-    env._close()
+    env.close()
 
 
 ray.init()
